Fix gear-test state logic
